refactor(schemas): reorganize auth-related schemas for scalability\n\n- Centralize Zod enums in src/shared/schemas/enums.ts derived from @shared/types/auth\n- Add shared primitives (UUID, ISODateTime, Email) in src/shared/schemas/primitives.ts\n- Move DB entity schemas to src/shared/schemas/entities/{user,org}.ts\n- Keep auth flow schemas in src/shared/schemas/auth.ts and import shared primitives/entities\n- Move org mgmt schemas to src/shared/schemas/orgs.ts\n- Move RBAC input schema to src/shared/schemas/roles.ts\n- Move HTTP response schemas to src/shared/schemas/http.ts\n- Update barrels and feature exports to reflect new structure\n\nNo functional behavior changes; only file structure and imports.
