+++ src/features/auth/index.ts
@@ -4,5 +4,9 @@
 export * from './server';
 export * from './utils';
 export * from '@shared/schemas/auth';
+export * from '@shared/schemas/roles';
+export * from '@shared/schemas/orgs';
+export * from '@shared/schemas/http';
+export * from '@shared/schemas/enums';
 export * from './roles';
 export { report } from '@shared/utils/error';
