+++ src/shared/schemas/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './user';
+export * from './org';
