+++ src/shared/schemas/index.ts
@@ -1,3 +1,9 @@
 // Unified schemas exports for the app
 export * from './auth';
 export * from './redirect';
+export * from './orgs';
+export * from './http';
+export * from './enums';
+export * from './roles';
+export * from './primitives';
+export * from './entities';
